Moved `api.py` to `http.py`
